refactor(snapshot): 管理页仿浏览记录Tabs重构为三Tab双列卡片
