+++ app/src/main/res/layout/fragment_snapshot_list.xml
@@ -1,0 +1,27 @@
+<?xml version="1.0" encoding="utf-8"?>
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:background="@color/v3_bg">
+
+    <androidx.recyclerview.widget.RecyclerView
+        android:id="@+id/snapshotList"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:clipToPadding="false"
+        android:padding="8dp"
+        tools:listitem="@layout/item_snapshot" />
+
+    <TextView
+        android:id="@+id/emptyHint"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_gravity="center"
+        android:gravity="center"
+        android:text="@string/snapshot_empty"
+        android:textColor="@color/v3_text_2"
+        android:textSize="15sp"
+        android:visibility="gone"
+        tools:visibility="visible" />
+</FrameLayout>
