--- app/src/main/res/layout/fragment_snapshot_manager.xml
@@ -1,58 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@color/v3_bg"
-    android:orientation="vertical">
-
-    <androidx.appcompat.widget.Toolbar
-        android:id="@+id/toolbar"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:background="?attr/colorPrimary"
-        android:elevation="@dimen/six_dp"
-        android:theme="@style/AppTheme.AppBarOverlay"
-        app:navigationIcon="@drawable/ic_arrow_back_white_shadow"
-        app:popupTheme="@style/AppTheme.PopupOverlay"
-        app:title="@string/snapshot_manager_title"
-        app:titleTextColor="@android:color/white" />
-
-    <FrameLayout
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_weight="1">
-
-        <androidx.recyclerview.widget.RecyclerView
-            android:id="@+id/snapshotList"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:clipToPadding="false"
-            android:padding="12dp"
-            tools:listitem="@layout/item_snapshot" />
-
-        <TextView
-            android:id="@+id/emptyHint"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center"
-            android:gravity="center"
-            android:text="@string/snapshot_empty"
-            android:textColor="@color/v3_text_2"
-            android:textSize="15sp"
-            android:visibility="gone"
-            tools:visibility="visible" />
-
-        <com.google.android.material.floatingactionbutton.FloatingActionButton
-            android:id="@+id/importButton"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="bottom|end"
-            android:layout_margin="16dp"
-            android:contentDescription="@string/snapshot_import"
-            android:src="@drawable/ic_baseline_get_app_24"
-            app:backgroundTint="?attr/colorPrimary"
-            app:tint="@android:color/white" />
-    </FrameLayout>
-</LinearLayout>
