+++ app/src/main/res/menu/menu_snapshot_manager.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+    <item
+        android:id="@+id/action_import_snapshot"
+        android:icon="@drawable/ic_baseline_get_app_24"
+        android:title="@string/snapshot_import"
+        app:showAsAction="always" />
+</menu>
